# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "e-Test API",
        "description": "Description",
        "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "security": [
        {
           "Bearer": []
        }
    ],
    "consumes": [],
    "produces": [],
    "paths": {
...
        "/api/v1/test": {
            "get": {
                "tags": ["test"],
                "summary": "Get all tests",
                "description": "",
                "parameters": [
                    {
                        "name": "select",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Test"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": ["test"],
                "summary": "Create test",
                "description": "",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Test"
                    }
                }],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Test"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/api/v1/section/": {
            "get": {
                "tags": ["section"],
                "summary": "Get all sections",
                "description": "",
                "parameters": [
                    {
                        "name": "select",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/responses/Section"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": ["section"],
                "summary": "Create section",
                "description": "",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Section"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/responses/Section"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/api/v1/candidate": {
            "get": {
                "tags": ["candidate"],
                "summary": "Get all candidates",
                "description": "",
                "parameters": [
                    {
                        "name": "select",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Candidate"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": ["candidate"],
                "summary": "Create candidate",
                "description": "",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Candidate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Candidate"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Candidate": {
            "required": [
                "firstName",
                "lastName",
                "phone",
                "email",
                "examType",
                "password"
            ],
            "properties": {
                "firstName": {
                    "type": "string",
                    "example": "Some example string"
                },
                "lastName": {
                    "type": "string",
                    "example": "Some example string"
                },
                "phone": {
                    "type": "string",
                    "example": "Some example string"
                },
                "email": {
                    "type": "string",
                    "example": "Some example string"
                },
                "examType": {
                    "type": "string",
                    "example": "Test id"
                },
                "password": {
                    "type": "string",
                    "example": "Some example string"
                },
                "createdAt": {
                    "type": "string",
                    "default": "Date.now"
                }
            }
        },
        "Test": {
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "example": "Some example string"
                },
                "timer": {
                    "type": "integer",
                    "default": 0
                },
                "createdAt": {
                    "type": "string",
                    "default": "Date.now"
                }
            }
        },
        "Section": {
            "required": [
                "title",
                "timer",
                "instruction",
                "test"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "example": "Some example string"
                },
                "timer": {
                    "type": "integer",
                    "default": 0
                },
                "instruction": {
                    "type": "string",
                    "example": "Some example string"
                },
                "test": {
                    "type": "string",
                    "example": "Test id"
                },
                "createdAt": {
                    "type": "string",
                    "default": "Date.now"
                }
            }
        },
...
    },
    "responses": {
        "Section": {
            "required": [
                "title",
                "timer",
                "instruction",
                "test"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "example": "Some example string"
                },
                "timer": {
                    "type": "integer",
                    "default": 0
                },
                "instruction": {
                    "type": "string",
                    "example": "Some example string"
                },
                "test": {
                    "$ref": "#/definitions/Test"                },
                "createdAt": {
                    "type": "string",
                    "default": "Date.now"
                }
            }
        },
...
    }
}